+++ src/julia-syntax.scm
@@ -1826,6 +1826,26 @@
                                  (call (top broadcasted) (top identity) ,e)))))))
 
 
+; Convert f(_,y) into x -> f(x,y) etcetera.  That is, an _ in e is changed into the
+; argument of an anonymous function.  Multiple underscores are turned into
+; multiple anonymous-function args.  2 arg-functions are turned into Fix1 or Fix2 calls.
+(define (curry-underscore e)
+  (expand-forms
+   (let ((oldargs (cddr e)))
+     (if (and (length= oldargs 2) (not (any kwarg? oldargs))) ; 2-arg case lowers to Fix1 or Fix2
+         (if (eq? '_ (car oldargs))
+             (if (eq? '_ (cadr oldargs))
+                 (cadr e) ; f(_,_) becomes f
+                 `(call (top Fix2) ,(cadr e) ,(cadr oldargs)))
+             `(call (top Fix1) ,(cadr e) ,(car oldargs)))
+         (let* ((args '()) ; n-arg case just becomes anon func
+                (enew (map (lambda (y) (if (eq? '_ y)
+                                           (let ((x (gensy)))
+                                             (set! args (cons x args))
+                                             x)
+                                           y)) e)))
+           `(-> (tuple ,@(reverse args)) ,enew))))))
+
 (define (expand-where body var)
   (let* ((bounds (analyze-typevar var))
          (v  (car bounds)))
@@ -2227,6 +2247,8 @@
                  ;; "(.op)(...)"
                  ((and (length= f 2) (eq? (car f) '|.|))
                   (expand-fuse-broadcast '() `(|.| ,(cadr f) (tuple ,@(cddr e)))))
+                 ((memq '_ (cddr e))
+                  (curry-underscore e))
                  ((eq? f 'ccall)
                   (if (not (length> e 4)) (error "too few arguments to ccall"))
                   (let* ((cconv (cadddr e))
@@ -2265,12 +2287,15 @@
                           (let ((x (car a)))
                             (if (and (length= x 2)
                                      (eq? (car x) '...))
-                                (if (null? run)
-                                    (list* (cadr x)
-                                           (tuple-wrap (cdr a) '()))
-                                    (list* `(call (core tuple) ,.(reverse run))
-                                           (cadr x)
-                                           (tuple-wrap (cdr a) '())))
+                                (begin
+                                  (if (eq? (cadr x) '_) ; _... is not currently allowed (meaning TBD)
+                                      (error (string "invalid underscore argument \"" (deparse x) "\"")))
+                                  (if (null? run)
+                                      (list* (cadr x)
+                                             (tuple-wrap (cdr a) '()))
+                                      (list* `(call (core tuple) ,.(reverse run))
+                                             (cadr x)
+                                             (tuple-wrap (cdr a) '()))))
                                 (tuple-wrap (cdr a) (cons x run))))))
                     (expand-forms
                      `(call (core _apply_iterate) (top iterate) ,f ,@(tuple-wrap argl '())))))